Address review: hook tests out of compliance suite, no docstring override
- Move the template-hook and carry-forward tests from the shipped
  ImageAPITest compliance suite into tests/test_rendering_hooks.py. The
  hook mechanism is an ImageViewerLogic implementation detail, not part
  of the API contract, so implementations must not be required to adopt
  it. The recording subclass no longer calls super() in the hooks, since
  they are no-ops in ImageViewerLogic.
- Drop the load_image docstring override so the docstring comes from the
  API definition; the carry-forward behavior remains described by inline
  comments and enforced by the moved tests.
- Revert the addition of image_viewer_logic to the package root
  namespace; nothing needs the root import and backends import from the
  module directly.
